fix(chat): Simplify MCP review fixes
Share queued Slack text combination between runtime paths and remove the redundant configured-provider check from the MCP manager API. Use activateProvider as the single provider activation gate so tool callers have fewer choices to misuse.
